feat: add content map generator
Built as a functional layer to create a global site map that stores all
of the various links arrays for rendering based on the content. This
will support inbound links, outbound links, resource links, etc. In
addition a content health file is also generated to assist in finding
any problematic files as the amount of content files grow.